+++ pkg/apis/platform/v1beta1/conditions.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2024-2025 ArangoDB GmbH, Cologne, Germany
+// Copyright 2024-2026 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -25,6 +25,7 @@ import api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 const (
 	ChartFoundCondition      api.ConditionType = "ChartFound"
 	DeploymentFoundCondition api.ConditionType = "DeploymentFound"
+	RouteFoundCondition      api.ConditionType = "RouteFound"
 	SpecValidCondition       api.ConditionType = "SpecValid"
 	ReleaseReadyCondition    api.ConditionType = "ReleaseReady"
 	ReadyCondition           api.ConditionType = "Ready"